Extract ImageSaveService, filewrapper for Preview, proj file
